renamed Poisson2D and Schrodinger2D
